sysusers: Provide meson argument to set gid for 'users' group
To allow better integration with distributions requiring an explicitly
set gid for the `users` group, provide the new `-Dusers-gid` option to
set to a new numeric value.

In the absence of a specified gid, we'll fallback to the default existing
behaviour of `-` as the gid value, to automatically assign the next available
gid on the system.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>