bpf-firewall: configure firewall ports and port ranges.
Introduce four new keywords, which have been modeled after
IPAddressAllow and IPAddressDeny keywords:

  IPPortIngressDeny and IPPortIngressAllow: which destination ports are
  denied or allowed for incoming TCP and UDP traffic.

  IPPortEgressDeny and IPPortEgressAllow: which destination ports are
  denied or allowed for outgoing TCP and UDP traffic.

The semantics work the same as with IPAddressAllow and IPAddressDeny. If
a port is both allowed and denied, it can be accessed. When used in
conjunction with IPAddressAllow and IPAddressDeny, a packet must pass
both IPAddress* and IPPort* tests in order to be accepted.