boot/efi: add support for automatic console mode switching
This patch adds support for setting the console mode to "auto". When
using this mode, systemd-boot will try to find a reasonable mode
automatically using some heuristic. Current implementation tries to
switch console mode to modes 2 if it exists, mode 1 if it exists, else
mode 0. Mode 2 is the first non standard mode provided by the firmware
developer and it's assumed to be a reasonable mode.

Usage: in systemd-boot configuration file loader/loader.conf add:

	config_mode auto